// api/request/user.request.ts
export class AddUserRequest extends BaseRequest {

    altrContactPhone: string;
    altrContactPhoneExtn: string;
    associatedAdminUsername: string;
    billingAce: string;
    billingAceList: string[];
    bpAddressList: string[];
    bpNameList: string[];
    bpNumberList: string[];
    brokerBPName: string;
    brokerBPNumber: string;
    contactPhone: string;
    contactPhoneExtn: string;
    emailId: string;
    faxNumber: string;
    firstName: string;
    hierarchyLevelList: string[];
    lastName: string;
    sapId: string;
    startStopAce: string;
    startStopAceList: string[];
    title: string;
    userName: string;
    webSecurityRole: string;
    linkTxnId:string;
}